fix: merged the createCustomer and authenticateCustomer files
