--- src/pages/login/index.ts
@@ -1,1 +1,0 @@
-export { authenticateCustomer } from "./authenticate-customer";
