--- src/pages/login/types.ts
@@ -1,4 +1,0 @@
-export interface BodyLogin {
-  email: string;
-  password: string;
-}
