--- src/pages/registration/index.ts
@@ -1,1 +1,0 @@
-export { createCustomer } from "./create-customer";
