+++ src/shared/api/index.ts
@@ -1,1 +1,3 @@
 export { ObtainAccessToken } from "./obtain-access-token";
+export { sendingSignInOrSignUpRequest } from "./signin-or-signup-request";
+export * from "./types";
