+++ src/shared/api/types.ts
@@ -1,3 +1,8 @@
+export interface BodyLogin {
+  email: string;
+  password: string;
+}
+
 interface Addresses {
   streetName: string;
   streetNumber: string;
@@ -14,4 +19,18 @@ export interface BodySignUp {
   addresses: Addresses[];
   dateOfBirth: string;
   store: string;
+}
+
+export interface AccessToken {
+  access_token: string;
+  expires_in: number;
+  refresh_token: string;
+  scope: string;
+  token_type: string;
+}
+
+export interface Error {
+  errors: [];
+  message: string;
+  statusCode: number;
 }
